pyproject.toml
[tool.poetry]
name = "dfr"
version = "0.1.0"
description = ""
authors = ["Nikita <nikita@mer1d1an.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2 = "^2.9.9"
black = "^24.4.2"
ipython = "^8.26.0"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
isort = "^5.13.2"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
